Draft: Split Shape2DView from Draft.py
